# lab4/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PizzGram Api Documentation",
        "description": "L5 Swagger PizzGram Api description",
        "contact": {
            "email": "eliseevv02@mail.ru"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "PizzGram API Server"
        }
    ],
    "paths": {
...
        "/checkouts": {
            "get": {
                "tags": [
                    "Checkouts"
                ],
                "summary": "Get list of all checkouts",
                "description": "Returns list of checkouts",
                "operationId": "getAllCheckouts",
                "responses": {
                    "200": {
                        "description": "success"
                    },
                    "401": {
                        "description": "Unauthorised"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Checkouts"
                ],
                "summary": "Add new checkout",
                "description": "Adds new checkout",
                "operationId": "addCheckout",
                "parameters": [
                    {
                        "name": "cart_id",
                        "in": "path",
                        "description": "Id of cart",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "checkout_state_id",
                        "in": "path",
                        "description": "Id of checkout state",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "sum",
                        "in": "path",
                        "description": "Final checkout cost",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "The adding request",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "cart_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "checkout_state_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "sum": {
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "The checkout was created.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "checkout": {
                                            "$ref": "#/components/schemas/Checkout"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorised"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/checkouts/{id}": {
            "get": {
                "tags": [
                    "Checkouts"
                ],
                "summary": "Get one checkout by id",
                "description": "Returns object of checkout",
                "operationId": "getCheckoutById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Checkout id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "checkout": {
                                            "$ref": "#/components/schemas/Checkout"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorised"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Checkouts"
                ],
                "summary": "Update existing checkout",
                "description": "Returns updated checkout data",
                "operationId": "updateCheckout",
                "parameters": [
                    {
                        "name": "cart_id",
                        "in": "path",
                        "description": "Id of cart",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "checkout_state_id",
                        "in": "path",
                        "description": "Id of checkout state",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "sum",
                        "in": "path",
                        "description": "Final checkout cost",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "The updating request",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "cart_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "checkout_state_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "sum": {
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "The checkout was updated.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Checkout"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Checkouts"
                ],
                "summary": "Delete checkout by id",
                "description": "Deletes checkout by id",
                "operationId": "deleteCheckout",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Checkout id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The checkout was deleted."
                    },
                    "401": {
                        "description": "Unauthorised"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Checkout": {
                "title": "Checkout",
                "description": "Checkout model",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "cart_id": {
                        "type": "integer"
                    },
                    "checkout_state_id": {
                        "type": "integer"
                    },
                    "sum": {
                        "type": "integer"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}